+++ src/lib/string.c
@@ -39,7 +39,7 @@ void *memset(void *dest, int c, size_t count)
     d = (uint8_t *)dest;
 
     while (count--) {
-        *d = c;
+        *d = (uint8_t)c;
         d++;
     }
 
@@ -53,5 +53,5 @@ size_t strnlen(const char *s, size_t n)
     for (str = s; *str != '\0' && n--; ++str) {
         /* Just iterate */
     }
-    return str - s;
+    return (size_t)((uintptr_t)str - (uintptr_t)s);
 }
